Fix lease parsing
- Fix parsing lease file when a uid contains a right brace
- Add kevent stubs so Linux compilation/testing works
