+++ go.mod
@@ -3,6 +3,7 @@ module github.com/nijave/go-dhcpleases
 go 1.16
 
 require (
-	github.com/nijave/go-dhcpd-leases v1.0.0
+	github.com/nijave/go-dhcpd-leases v1.0.2
 	github.com/sirupsen/logrus v1.8.1
+	golang.org/x/sys v0.0.0-20220330033206-e17cdc41300f // indirect
 )
