+++ plugin/Makefile
@@ -1,5 +1,5 @@
 PLUGIN_NAME=		go-dhcpleases
-PLUGIN_VERSION=		1.2
+PLUGIN_VERSION=		1.3
 PLUGIN_COMMENT=		dhcpleases reimplementation
 PLUGIN_DEPENDS=
 PLUGIN_MAINTAINER=	nick@venenga.com
